refactor(ava): move advise into a floder
